feat-[166272346]: Users follow each other move the follow and unfollow controller from user controller to profile controller update the follow and unfollow method to return profile of who followed
